Se agrega Stored Procedure en listado de Profesionales

+++ dominio/HorarioProfesional.cs
@@ -7,11 +7,11 @@ using System.Threading.Tasks;
 namespace dominio
 {
     public class HorarioProfesional
-    {   
+    {
         public Profesional Profesional { get; set; }
         public Consultorio Consultorio { get; set; }
-        public int DiaAtencion {  get; set; }
+        public byte DiaAtencion { get; set; }
         public TimeSpan HoraInicio { get; set; }
-        public TimeSpan HoraFin {get; set; }
+        public TimeSpan HoraFin { get; set; }
     }
 }
